Add missing features
