--- crates/json-rpc/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "json-rpc"
-version = "0.1.0"
-edition = "2021"
-
-[dev-dependencies]
-tokio = { workspace = true, features = ["full"] }
-
-[dependencies]
-futures = { workspace = true }
-hyper = "0.14.27"
-jsonrpsee = { version = "0.22", features = ["server"] }
-reqwest = { version = "0.12", features = ["json"] }
-serde = { workspace = true, features = ["derive", "rc"] }
-serde_json = { workspace = true }
-tower = "0.4.13"
-tower-http = { version = "0.4.4", features = ["cors"] }
-url = "2.5"
