+++ crates/json-rpc/json-rpc-client/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "json-rpc-client"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+futures = { workspace = true }
+reqwest = { version = "0.12", features = ["json"] }
+serde = { workspace = true, features = ["derive", "rc"] }
+serde_json = { workspace = true }
