--- crates/json-rpc/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod server;
